Shells/jmeter/src/requirements.txt
cloudshell-shell-core>=3.1,<4.0
cloudshell-core
cloudshell-automation-api
requests
paramiko
pygithub